Add all OMP data motion
Now we allocate data on the device using the OpenMP
target data region and manually control the data motion.
This will turn off all of the automatic large data motions.
However, OMP is still moving data automatically that was not
in the map(alloc:) statement. This should be mostly small
scalar data at this point.

There's also two target offload regions in the CG algorithm
here that did not work correctly without the data movement.
I did not dig deeper as to why this was the case.